fixup: clean up files
